W141: strip stale markers for mdlFile allocator-provenance triplet 003398C0/0033B1C0/0033E7F0

+++ src/Event/Fcl/fclCombineMisc.c
@@ -51,7 +51,7 @@ s32 FUN_003d6e60(int param_1,int param_2);
 u32 FUN_003d6f80(int param_1,int param_2,void *param_3);
 u32 FUN_003d71d0(int param_1,int param_2,void *param_3);
 u64 FUN_003d72f0(u16 *param_1);
-void FUN_003d74f0(long param_1,long param_2,int param_3);
+void FUN_003d74f0(int param_1,int param_2,int param_3);
 #pragma alias FUN_003d74f0_i FUN_003d74f0
 extern void FUN_003d74f0_i(int param_1, int param_2, int param_3);
 u32 FUN_003d7ac0(char param_1,u8 param_2,u16 *param_3);
@@ -1791,7 +1791,7 @@ LAB_003d7430:
 // FUN_003D74F0 NONMATCHING
 
 
-void FUN_003d74f0(long param_1,long param_2,int param_3)
+void FUN_003d74f0(int param_1,int param_2,int param_3)
